DP 3333h Find the Original Typed String II
    Finished `test` profile [unoptimized + debuginfo] target(s) in 0.76s
     Running unittests src/lib.rs (target/debug/deps/dp-8ed43879319390e2)

running 1 test
* "aabbccdd" 7
-> [2, 2, 2, 2]
:: 5
* "aabbccdd" 8
-> [2, 2, 2, 2]
:: 1
* "aaabbb" 3
-> [3, 3]
:: 8
* "ggggggggaaaaallsssssaaaaaaaaaiiqqqqqqqqqqbbbbbbbvvfffffjjjjeeeeeefffmmiiiix" 34
-> [8, 5, 2, 5, 9, 2, 10, 7, 2, 5, 4, 6, 3, 2, 4, 1]
:: 834168507
test tests::test_3333 ... ok

test result: ok. 1 passed; 0 failed; 0 ignored; 0 measured; 33 filtered out; finished in 0.00s

+++ dp/src/lib.rs
@@ -1436,7 +1436,7 @@ impl Sol3333 {
         counts[0][0] = 1;
 
         for g in 0..freqs.len() {
-            for l in 1..k as usize {
+            for l in 1..k {
                 for x in 1..=freqs[g] as usize {
                     if l >= x {
                         counts[g + 1][l] = (counts[g + 1][l] + counts[g][l - x]) % M;
